requirement.txt
opencv-python==4.8.1.78
face-recognition==1.3.0
mysql-connector-python==8.1.0
supabase==1.0.4
numpy==1.24.3
python-dotenv==1.0.0
schedule==1.2.0
pillow==10.0.1
requests==2.31.0